Create Stock buy and sell II
You are given an integer array prices where prices[i] is the price of a given stock on the ith day.

On each day, you may decide to buy and/or sell the stock. You can only hold at most one share of the stock at any time. However, you can buy it then immediately sell it on the same day.

Find and return the maximum profit you can achieve.

Approach:

Think of this as collecting profits from every small increase in stock prices.

Scan the array and whenever prices[i] > prices[i-1], add the difference to the total profit.
If the price drops or stays the same, do nothing and move to the next day.